leo-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "leo-core"
version = "0.2.0"
description = "LEO Core — AI Visibility Scoring Engine using LangGraph-style Agents"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Yeshwanth Sai", email = "founder@tekyantra.com"}]
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "typer>=0.12.5",
    "pydantic>=2.9.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.0",
    "psycopg2-binary>=2.9.0",
    "openai>=1.52.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "black>=24.0", "flake8>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["leo*", "api*"]

[project.urls]
Homepage = "https://github.com/Yesh48/LEOlabs"
Issues = "https://github.com/Yesh48/LEOlabs/issues"
